# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "My test service",
        "title": "My test service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Artem Rotov",
            "url": "https://github.com/ArtemRotov",
            "email": "rotoffff@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/account/": {
            "get": {
                "description": "User balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/account"
                ],
                "summary": "Balance",
                "parameters": [
                    {
                        "description": "user_id",
                        "name": "balanceInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.balanceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.balanceOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/account/deposit/": {
            "post": {
                "description": "Deposit by userId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/account/deposit"
                ],
                "summary": "Deposit",
                "parameters": [
                    {
                        "description": "user_id, amount",
                        "name": "depositInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.depositInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.depositOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/create": {
            "post": {
                "description": "create new reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/reservation/create"
                ],
                "summary": "create",
                "parameters": [
                    {
                        "description": "ID NO NEED",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ArtemRotov_account-balance-manager_internal_model.Reservation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.createOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "Sign in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "ID NO NEED",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ArtemRotov_account-balance-manager_internal_model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.signInOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Sign up",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "ID NO NEED",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ArtemRotov_account-balance-manager_internal_model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.signUpOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorOutput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ArtemRotov_account-balance-manager_internal_model.Reservation": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 12
                },
                "amount": {
                    "type": "integer",
                    "example": 12441
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer",
                    "example": 11231
                },
                "service_id": {
                    "type": "integer",
                    "example": 134
                }
            }
        },
        "github_com_ArtemRotov_account-balance-manager_internal_model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 6,
                    "example": "pass12345678"
                },
                "username": {
                    "type": "string",
                    "example": "example@mail.org"
                }
            }
        },
        "internal_controller_http_v1.ErrorOutput": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "example error"
                }
            }
        },
        "internal_controller_http_v1.balanceInput": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_controller_http_v1.balanceOutput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 10000
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_controller_http_v1.createOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_controller_http_v1.depositInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_controller_http_v1.depositOutput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 1000
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_controller_http_v1.signInOutput": {
            "type": "object",
            "properties": {
                "Token": {
                    "type": "string",
                    "example": "eyJhbGc..."
                }
            }
        },
        "internal_controller_http_v1.signUpOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}